The first regression round was red, and it was right twice
CI's `stable` is 1.98.1, released three days ago; this machine is on
1.96.0. Two versions of new clippy lints, all fatal under -D warnings,
and the round found 47 of them. Reproduced by installing 1.98.1 beside
the default toolchain rather than guessing, since Actions logs need a
token to read.

Forty-three are one lint — `chunks_exact_to_as_chunks` — on this
codebase's RGBA pixel idiom. `as_chunks::<4>()` returns a tuple and
changes the shape of every one of those call sites, in the code where a
slip is a rendering bug, for nothing at runtime. The lint is right in
general and wrong here, so the workspace says so once, in a
`[workspace.lints.clippy]` table every crate inherits. promo-gpu repeats
the line instead: Cargo refuses a crate that both inherits and declares,
and it already declares `unexpected_cfgs` for objc's macro output.

The other four are real, and two of them are mine from this morning — a
`mut` left behind when the reader stamp stopped being assigned, and a
`u32 as u32` in a test. The older toolchain here never saw either. That
is the argument for leaving CI on `stable` rather than pinning it to
whatever this Mac happens to have: a round that only ever runs the
toolchain that wrote the code cannot tell you anything the author did not
already know.

Also two `?` rewrites clippy suggested and I took, in `layout_check`'s
sprite branch and `preview`'s drawing branch.

Green now under both toolchains: fmt, clippy -D warnings, 23 suites.

+++ promo-text/src/lib.rs
@@ -1523,7 +1523,7 @@ mod tests {
                 .expect("rasterized");
             // The lit span of the FIRST line only — the short one.
             let (mut left, mut right) = (out.width as i64, 0i64);
-            let half = (out.height / 2) as u32;
+            let half = out.height / 2;
             for y in 0..half {
                 for x in 0..out.width {
                     if out.rgba[((y * out.width + x) * 4 + 3) as usize] > 8 {
